added new bigDecimal scaffold page

+++ .g8/bigDecimalPage/app/pages/$className$Page.scala
@@ -2,7 +2,7 @@ package pages
 
 import play.api.libs.json.JsPath
 
-case object $className$Page extends QuestionPage[Int] {
+case object $className$Page extends QuestionPage[BigDecimal] {
   
   override def path: JsPath = JsPath \ toString
